Add entity types label mapping (#21) Also move "common" out of "pages"
