+++ src/common/entityTypes.js
@@ -1,0 +1,15 @@
+// Entity type description labels
+export const entityTypes = {
+    attendance: "Beckett's attendance of events",
+    music: "Specific musical works",
+    organization: "Entities with corporate purpose",
+    person: "Persons named",
+    place: "Geographical locations",
+    production: "Performances mentioned",
+    public_event: "Significant world events",
+    publication: "Process and published product",
+    reading: "Any text received/read/mentioned",
+    translating: "Mentions of translation(s)",
+    work_of_art: "Specific works/monuments",
+    writing: "Mentions of writing practice/context",
+};
