+++ src/common/index.js
@@ -1,4 +1,5 @@
 export * from "./dateUtil";
+export * from "./entityTypes";
 export * from "./fetch";
 export * from "./queryBuilder";
 export * from "./volumeLabels";
