+++ src/common/result.css
