+++ src/pages/LettersSearchPage/LettersSearchPage.jsx
@@ -40,7 +40,7 @@ import ValueFilter from "../../components/ValueFilter";
 import DateRangeFacet from "../../components/DateRangeFacet";
 import DateRangeFilter from "../../components/DateRangeFilter";
 import withSearchRouting from "../../components/withSearchRouting";
-import "../common/search.css";
+import "../../common/search.css";
 
 // icon component cache required for dynamically imported EUI icons in Vite;
 // see https://github.com/elastic/eui/issues/5463
